+++ json-schema/src/main/kotlin/io/bkbn/kompendium/json/schema/handler/SimpleObjectHandler.kt
@@ -1,6 +1,7 @@
 package io.bkbn.kompendium.json.schema.handler
 
 import io.bkbn.kompendium.json.schema.SchemaGenerator
+import io.bkbn.kompendium.json.schema.SerializableReader
 import io.bkbn.kompendium.json.schema.definition.JsonSchema
 import io.bkbn.kompendium.json.schema.definition.NullableDefinition
 import io.bkbn.kompendium.json.schema.definition.OneOfDefinition
@@ -24,18 +25,24 @@ import kotlin.reflect.jvm.javaField
 
 object SimpleObjectHandler {
 
-  fun handle(type: KType, clazz: KClass<*>, cache: MutableMap<String, JsonSchema>): JsonSchema {
+  fun handle(
+    type: KType,
+    clazz: KClass<*>,
+    cache: MutableMap<String, JsonSchema>,
+    serializableReader: SerializableReader
+  ): JsonSchema {
 
     cache[type.getSimpleSlug()] = ReferenceDefinition(type.getReferenceSlug())
 
     val typeMap = clazz.typeParameters.zip(type.arguments).toMap()
-    val props = clazz.serializableMemberProperties()
+    val props = serializableReader.serializableMemberProperties(clazz)
+      .filterNot { it.javaField == null }
       .associate { prop ->
       val schema = when (prop.needsToInjectGenerics(typeMap)) {
-        true -> handleNestedGenerics(typeMap, prop, cache)
+        true -> handleNestedGenerics(typeMap, prop, cache, serializableReader)
         false -> when (typeMap.containsKey(prop.returnType.classifier)) {
-          true -> handleGenericProperty(prop, typeMap, cache)
-          false -> handleProperty(prop, cache)
+          true -> handleGenericProperty(prop, typeMap, cache, serializableReader)
+          false -> handleProperty(prop, cache, serializableReader)
         }
       }
 
@@ -44,13 +51,15 @@ object SimpleObjectHandler {
         false -> schema
       }
 
-      prop.serializableName() to nullCheckSchema
+      serializableReader.serializableName(prop) to nullCheckSchema
     }
 
-    val required = clazz.serializableMemberProperties()
+    val required = serializableReader.serializableMemberProperties(clazz)
+      .asSequence()
+      .filterNot { it.javaField == null }
       .filterNot { prop -> prop.returnType.isMarkedNullable }
       .filterNot { prop -> clazz.primaryConstructor!!.parameters.find { it.name == prop.name }!!.isOptional }
-      .map { it.serializableName() }
+      .map { serializableReader.serializableName(it) }
       .toSet()
 
     val definition = TypeDefinition(
@@ -65,16 +74,6 @@ object SimpleObjectHandler {
     }
   }
 
-  private fun KClass<*>.serializableMemberProperties() =
-    memberProperties
-      .filterNot { it.hasAnnotation<Transient>() }
-      .filterNot { it.javaField == null }
-
-  private fun KProperty1<out Any, *>.serializableName() =
-    annotations
-      .filterIsInstance<SerialName>()
-      .firstOrNull()?.value?: name
-
   private fun KProperty<*>.needsToInjectGenerics(
     typeMap: Map<KTypeParameter, KTypeProjection>
   ): Boolean {
@@ -85,7 +84,8 @@ object SimpleObjectHandler {
   private fun handleNestedGenerics(
     typeMap: Map<KTypeParameter, KTypeProjection>,
     prop: KProperty<*>,
-    cache: MutableMap<String, JsonSchema>
+    cache: MutableMap<String, JsonSchema>,
+    serializableReader: SerializableReader
   ): JsonSchema {
     val propClass = prop.returnType.classifier as KClass<*>
     val types = prop.returnType.arguments.map {
@@ -93,7 +93,7 @@ object SimpleObjectHandler {
       typeMap.filterKeys { k -> k.name == typeSymbol }.values.first()
     }
     val constructedType = propClass.createType(types)
-    return SchemaGenerator.fromTypeToSchema(constructedType, cache).let {
+    return SchemaGenerator.fromTypeToSchema(constructedType, cache, serializableReader).let {
       if (it.isOrContainsObjectDef()) {
         cache[constructedType.getSimpleSlug()] = it
         ReferenceDefinition(prop.returnType.getReferenceSlug())
@@ -106,11 +106,12 @@ object SimpleObjectHandler {
   private fun handleGenericProperty(
     prop: KProperty<*>,
     typeMap: Map<KTypeParameter, KTypeProjection>,
-    cache: MutableMap<String, JsonSchema>
+    cache: MutableMap<String, JsonSchema>,
+    serializableReader: SerializableReader
   ): JsonSchema {
     val type = typeMap[prop.returnType.classifier]?.type
       ?: error("This indicates a bug in Kompendium, please open a GitHub issue")
-    return SchemaGenerator.fromTypeToSchema(type, cache).let {
+    return SchemaGenerator.fromTypeToSchema(type, cache, serializableReader).let {
       if (it.isOrContainsObjectDef()) {
         cache[type.getSimpleSlug()] = it
         ReferenceDefinition(type.getReferenceSlug())
@@ -120,8 +121,12 @@ object SimpleObjectHandler {
     }
   }
 
-  private fun handleProperty(prop: KProperty<*>, cache: MutableMap<String, JsonSchema>): JsonSchema =
-    SchemaGenerator.fromTypeToSchema(prop.returnType, cache).let {
+  private fun handleProperty(
+    prop: KProperty<*>,
+    cache: MutableMap<String, JsonSchema>,
+    serializableReader: SerializableReader
+  ): JsonSchema =
+    SchemaGenerator.fromTypeToSchema(prop.returnType, cache, serializableReader).let {
       if (it.isOrContainsObjectDef()) {
         cache[prop.returnType.getSimpleSlug()] = it
         ReferenceDefinition(prop.returnType.getReferenceSlug())